Move the abort report out of the header every unit includes
