+++ include/tts/engine/abort.hpp
@@ -9,7 +9,6 @@
 #pragma once
 
 #include <tts/engine/deps.hpp>
-#include <tts/engine/environment.hpp>
 #include <tts/tools/callable.hpp>
 #include <tts/tools/erased_storage.hpp>
 
@@ -67,26 +66,8 @@ namespace tts::_
     }
   };
 
-  inline callable      abort_epilogue  = {}; // NOSONAR - the driver sets it once it is ready
-  inline abort_handler abort_action    = {}; // NOSONAR - set_abort_handler replaces it
-
-  inline std::size_t   remaining_tests = 0; // NOSONAR - the loop updates it before each case
-
-  // What a dying run tells the sinks. The crash guard and the watchdog differ by two words.
-  inline void report_abort(char const* headline, char const* reason)
-  {
-    // Without fatal() the Results: line reads 100% success on a run that died.
-    ::tts::global_runtime.fatal();
-    ::tts::global_runtime.unexpected();
-
-    ::tts::output().writeln(headline);
-    ::tts::output().suite_aborted();
-    ::tts::output().writeln(
-    "@@ ABORTING DUE TO %s @@ - %d Tests not run", reason, static_cast<int>(remaining_tests));
-
-    ::tts::report(0, 0);
-    ::tts::output().finish();
-  }
+  inline callable          abort_epilogue = {}; // NOSONAR - the driver sets it once it is ready
+  inline abort_handler     abort_action   = {}; // NOSONAR - set_abort_handler replaces it
 
   [[noreturn]] inline void exit_now()
   {
